Import from the internal repository.
git-svn-id: https://mozc.googlecode.com/svn/trunk@49 a6090854-d499-a067-5803-1114d4e51264

+++ src/session/session_converter_test.cc
@@ -1528,7 +1528,7 @@ TEST_F(SessionConverterTest, FillContext) {
       "\xE8\xBB\x8A\xE3\x81\xA7",
       "\xE8\xA1\x8C\xE3\x81\x8F"
   };
-  for (int i = 0; i < arraysize(kHistoryInput); ++i) {
+  for (size_t i = 0; i < arraysize(kHistoryInput); ++i) {
     Segment *segment = segments.add_segment();
     segment->set_segment_type(Segment::HISTORY);
     Segment::Candidate *candidate = segment->add_candidate();
